+++ map.py
@@ -1,3 +1,39 @@
+import json
+
+
+class Map:
+    def __init__(self, path) -> None:
+        try:
+            with open(path, "r", encoding="utf-8", errors="replace") as file:
+                contents = file.read()
+            objects = [json.loads(s.strip()) for s in contents.splitlines()]
+            # Extract gold, fluxite, and artifacts from the first JSON segment at path resources.gold, fluxite, artifacts
+            resources = objects[0].get("resources", {}) if objects else {}
+            self.gold = resources.get("gold", 0)
+            self.fluxite = resources.get("fluxite", 0)
+            self.artifacts = resources.get("artifacts", 0)
+
+            found = False
+            for obj in objects:
+                if "world" in obj and "matrix" in obj["world"]:
+                    found = True
+                    self.world = obj["world"]["matrix"]
+                    break
+            if not found:
+                print("Could not find world.matrix in save file.")
+                return
+            if not (isinstance(self.world, list) and all(isinstance(row, list) for row in self.world)):
+                print("world.matrix is not a valid tilemap.")
+                return
+            player_data = objects[1].get("player", {}) if len(objects) > 1 else {}
+            self.player_x = player_data.get("x", 0) / 4
+            self.player_y = player_data.get("y", 0) / 4
+            self.active_slot = player_data.get("activeslotindex", 0)
+        except Exception as e:
+            print(f"Error loading tilemap: {e}")
+            return
+
+
 class TileInfo:
     def __init__(self, color: tuple[int, int, int], name: str, hex_code: str):
         self.color = color
